pyproject.toml
[tool.poetry]
name = "eva-auth"
version = "0.1.0"
description = "Authentication and authorization service for EVA Suite"
authors = ["EVA Team <eva@example.com>"]
readme = "README.md"
packages = [{include = "eva_auth", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
PyJWT = "^2.8.0"
authlib = "^1.3.0"
httpx = "^0.26.0"
redis = "^5.0.1"
azure-identity = "^1.15.0"
azure-keyvault-secrets = "^4.7.0"
azure-cosmos = "^4.5.1"
cryptography = "^41.0.0"
python-dotenv = "^1.0.0"
structlog = "^24.1.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
fakeredis = "^2.21.0"
httpx = "^0.26.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
black = "^23.12.0"
isort = "^5.13.0"
pre-commit = "^3.6.0"
locust = "^2.42.6"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=eva_auth",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=100"
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "DTZ", "T10", "RET", "SIM", "ARG"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
